Typedef cleansing - u_long to unsigned int, u_short to unsigned short.
As before, valid because these date from the 32-bit era. Resolves a very
old to-do item.

+++ include/ntp_stdlib.h
@@ -59,7 +59,7 @@ extern	int	authencrypt	(keyid_t, uint32_t *, int);
 extern	int	authhavekey	(keyid_t);
 extern	int	authistrusted	(keyid_t);
 extern	bool	authreadkeys	(const char *);
-extern	void	authtrust	(keyid_t, unsigned long);
+extern	void	authtrust	(keyid_t, unsigned int);
 extern	bool	authusekey	(keyid_t, int, const uint8_t *);
 
 /*
